requirements.txt
pytenable
pan-os-python
psycopg2-binary
pandas
requests
flask
flask-cors
python-dotenv
python-nmap
Faker
